requirements.txt
flask
flask-cors
paho-mqtt
numpy
